Centralising search type result types and allow merging declarations.
