+++ graylog2-web-interface/src/views/types.d.ts
@@ -38,6 +38,7 @@ import {
   WidgetConfigFormValues,
 } from 'views/components/aggregationwizard';
 import VisualizationConfig from 'views/logic/aggregationbuilder/visualizations/VisualizationConfig';
+import { TimeRange } from 'views/logic/queries/Query';
 
 interface EditWidgetComponentProps<Config extends WidgetConfig = WidgetConfig> {
   children: React.ReactNode,
@@ -162,6 +163,23 @@ interface SystemConfiguration {
     updateConfig: (newConfig: any) => any,
   }>;
 }
+
+export type SearchTypeResult = {
+  type: string,
+  effective_timerange: TimeRange,
+};
+
+export type MessageResult = {
+  type: 'messages',
+  total: number,
+};
+
+export interface SearchTypeResultTypes {
+  generic: SearchTypeResult,
+  messages: MessageResult,
+}
+
+export type SearchTypeResults = { [id: string]: SearchTypeResultTypes[keyof SearchTypeResultTypes] };
 
 declare module 'graylog-web-plugin/plugin' {
   export interface PluginExports {
